更改时间显示问题

+++ qinjee-parent/qinjee-masterdata/src/main/java/com/qinjee/masterdata/model/entity/UserArchive.java
@@ -124,7 +124,7 @@ public class UserArchive implements Serializable {
      * 年龄
      */
     @ApiModelProperty("年龄")
-    private Short age;
+    private Integer age;
 
     /**
      * 民族
